Add test cases and GitHub action workflow
This change adds test cases for most methods in debuginfod.py.

For now, this test case requires a Fedora base system as we
can only download source and executable files from the Fedora
server. We also use a pinned version of the gcc package to
reproducibly get a source file. The test cases run as a
GitHub Action and pull a Fedora Rawhide container to have a
working libdebuginfod.so installation from elfutils 0.183.